+++ test/testing/SmallDatabaseImpl.java
@@ -54,7 +54,7 @@ public class SmallDatabaseImpl implements Database {
 
 
   @Override
-  public void logEvent(String exid, String context, long userid, String device) {
+  public void logEvent(String exid, String context, int userid, String device) {
 
   }
